Restructured java projects. Implemented communication between java and cpp. Started working on java UI.

+++ FerChessAI/Sources/Math.h
@@ -18,7 +18,7 @@ inline int AbsI(int Value)
 	return Value >= 0 ? Value : -Value;
 }
 
-inline int AbsF(float Value)
+inline float AbsF(float Value)
 {
 	return Value >= 0.0f ? Value : -Value;
 }
